django_app/account/urls.py
from django.urls import path
from .views import CreateUserAPIView, CustomTokenObtainPairView
from rest_framework_simplejwt.views import TokenRefreshView

urlpatterns = [
    path('login/', CustomTokenObtainPairView.as_view(), name='custom_token_obtain_pair'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('user-create/', CreateUserAPIView.as_view(), name='create_user'),
]