requirements.txt
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
redis==5.2.0
elasticsearch==7.10.0
djangorestframework-simplejwt==5.2.0
